Move isExecutableSync() to a focused module.
